restructure endpoint POST /company to unit function suitable for testing
